+++ CHANGELOG.md
@@ -5,7 +5,9 @@
     This means that `Entity :: has ()` returns `true` even for properties with
     empty or null value;
 * added `Entity::hasValue()` and `Entity::isEmpty()` methods;
-* added `array_to_string()` global function.
+* added `array_to_string()` global function;
+* `is_stringable()` function now returns `true` for arrays that can be converted
+    to strings with `array_to_string ()`.
 
 ### 1.5.7
 * `deprecationWarning()` global function has been moved to `src/deprecation_functions.php`.